Robot commands now return a status, trajectories now use milliseconds

+++ include/spot/trajectory.h
@@ -31,12 +31,13 @@ class Trajectory3D{
 public:
     Trajectory3D();
 
-    void addPointRPY(double x, double y, double z, double roll, double pitch, double yaw, double time);
-	void addPointRPYVel(double x, double y, double z, double roll, double pitch, double yaw, double time,
+    void addPointRPY(double x, double y, double z, double roll, double pitch, double yaw, int64_t time);
+	void addPointRPYVel(double x, double y, double z, double roll, double pitch, double yaw, int64_t time,
 		    double velX, double velY, double velZ, double angVelX, double angVelY, double angVelZ);
-	void addPointQuat(double x, double y, double z, double qx, double qy, double qz, double qw, double time);
-	void addPointQuatVel(double x, double y, double z, double qx, double qy, double qz, double qw, double time,
+	void addPointQuat(double x, double y, double z, double qx, double qy, double qz, double qw, int64_t time);
+	void addPointQuatVel(double x, double y, double z, double qx, double qy, double qz, double qw, int64_t time,
 		    double velX, double velY, double velZ, double angVelX, double angVelY, double angVelZ);
+
     void setPosInterp(bool cubic);
 	void setAngInterp(bool cubic);
